==> JTHReplication/1-ReplicationSample.ipynb <==
import numpy as np
import scipy as sp
import pandas as pd
import timeit
import datetime
import time
import pprint
import itertools
import pickle
import sklearn
import dask.dataframe as dd
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
import fastparquet
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
# Distances
import scipy.spatial.distance as distance
# KL
from scipy.stats import entropy
# Normalize
from sklearn.preprocessing import normalize
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
# Pairwise distances
from sklearn.metrics.pairwise import pairwise_distances
import h5py

Code taken from:
- [`201801KnowledgeSpilloversRep/Replication-1-JTH.ipynb`](https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201801KnowledgeSpilloversRep/Replication-1-JTH.ipynb)
- [`201801KnowledgeSpilloversRep/Replication-2-ControlSelection.ipynb`](https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201801KnowledgeSpilloversRep/Replication-2-ControlSelection.ipynb)
- [`201801KnowledgeSpilloversRep/Replication-3-CitationSimilarity.ipynb`](https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201801KnowledgeSpilloversRep/Replication-3-CitationSimilarity.ipynb)
- [`201801KnowledgeSpilloversRep/Replication-4-SummaryTables.ipynb`](https://sfengc7.stern.nyu.edu:8888/notebooks/patentdiffusion/201801KnowledgeSpilloversRep/Replication-4-SummaryTables.ipynb)

pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq").to_pandas()
print(len(pdf))
pat_dup = pd.read_pickle("RawData/Cleaned/duplicate_pattext_0712.pkl")
pdf = pdf.loc[~(pdf["patent"].isin(pat_dup))]
print(len(pdf))
# Check I have abstracts
pa = fastparquet.ParquetFile("RawData/Cleaned/patabs7615_us_no_dup.parq").to_pandas(["patent"])["patent"]
pdf = pdf.loc[pdf["patent"].isin(pa)]
print(len(pdf))

# Only use patents that have been cited
cit = dd.read_parquet("RawData/Cleaned/cit_0628.parq")
cit = cit["cited"]

pdf = pdf.loc[pdf["patent"].isin(cit.compute())]
print(len(pdf))
del(cit)

### Sampling procedure

1. Group patents by primary class; get index
2. Get slice of patents' indices
3. Get slide of patents' indices+1

Removal rules
- Remove pairs from different grant years
- Remove pairs from same location
- Remove pairs from same assignee

# No missing
pdf = pdf.loc[pdf["primclass"].notnull()]
# Exclude primary classes with less than 100 patents
vc = pdf["primclass"].value_counts()
vc = vc[vc > 100].index.tolist()
pdf = pdf.loc[pdf["primclass"].isin(vc)]
print(len(pdf))

%%time
pg = {n: np.array(g["patent"].tolist()) for n,g in pdf[["patent", "primclass"]].groupby("primclass")}
pgi = {n: np.array(range(len(g))) for n,g in pg.items()}

%%time

def samp_ind(g):
    
    if len(g)>10000:
        size = int(np.round(0.60*len(g)))
        t = np.random.choice(g, size = size, replace = False)
    else:
        size = int(np.round(0.8*len(g)))
        t = np.random.choice(g, size = size, replace = False)
    
    # Remove any indices greater than the length of the entire array
    t = t[np.where(t<len(g)-2)]
    
    c = t+1
    return t,c

# Sample
ts = pd.DataFrame()
for n,g in pgi.items():
    s = pd.DataFrame()
    tpi, cpi = samp_ind(g)
    s["tp"] = pg[n][tpi]
    s["cp"] = pg[n][cpi]
    s["tp_primclass"] = n
    ts = ts.append(s, ignore_index = True)

# Grant year and MSA
ts = ts.merge(pdf[["gyear", "inv_msa", "patent"]].add_prefix("tp_"), how = "left", left_on = "tp",
             right_on = "tp_patent").drop("tp_patent",1)
ts = ts.merge(pdf[["gyear", "inv_msa", "patent"]].add_prefix("cp_"), how = "left", left_on = "cp",
             right_on = "cp_patent").drop("cp_patent",1)

# Remove non-matching grant year
print(len(ts))
ts = ts.loc[ts["tp_gyear"] == ts["cp_gyear"]]
print(len(ts))

# Remove matching MSA
ts = ts.loc[ts["tp_inv_msa"] != ts["cp_inv_msa"]]
print(len(ts))

# %%time
# asgs = fastparquet.ParquetFile("RawData/Cleaned/patent_assignees_0628.parq").to_pandas(["patent", "assignee_id"])
# asgs = {n: g["assignee_id"].tolist() for n,g in asgs.groupby("patent")}

# # Save assignee dictionary
# pickle.dump(asgs, open("RawData/Cleaned/patent_assignee_dict_0628.pkl", "wb"))

%%time
asgs = pickle.load(open("RawData/Cleaned/patent_assignee_dict_0628.pkl", "rb"))

# Check that target and control are in assignee list
%time asg_match = (set(asgs.get(tp, [])).intersection(asgs.get(cp, [])) for tp, cp in zip(ts["tp"], ts["cp"]))
%time asg_match = [len(i) for i in asg_match]
del(asgs)

ts["asg_match"] = asg_match
ts = ts.loc[ts["asg_match"] == 0]
ts = ts.drop("asg_match",1)
print(len(ts))

ts.head()

ts.to_pickle("DataStore/2018-07-P2/targ_samp_0717.pkl")

______
## Citations from Sample

1. Get citations by each target and control patent
2. Remove self-citations
3. Group by target and control
4. Get percentage match to *tp_inv_msa*
5. Get list of *tp*'s citations that match *tp_inv_msa*

ts = pd.read_pickle("DataStore/2018-07-P2/targ_samp_0717.pkl")
tpats = list(ts["tp"])+list(ts["cp"])

cit = dd.read_parquet("RawData/Cleaned/cit_0628.parq")
cit = cit[cit["cited"].isin(tpats)]

# Convert back to pandas
cit = cit.compute()
print(len(cit))

%%time
# Remove self citations
asgs = pickle.load(open("RawData/Cleaned/patent_assignee_dict_0628.pkl", "rb"))
%time asg_match = (set(asgs.get(cited, [])).intersection(asgs.get(citing, [])) for cited, citing \
                   in zip(cit["cited"], cit["citing"]))
%time asg_match = [len(i) for i in asg_match]
del(asgs)

cit["asg_match"] = asg_match
cit = cit.loc[cit["asg_match"] == 0]
cit = cit[["citing", "cited"]]
print(len(cit))

# Add tp to each citation
tdict = dict(zip(ts["tp"], ts["tp"]))
tdict.update(dict(zip(ts["cp"], ts["tp"])))
cit["tp"] = cit["cited"].map(tdict)
del(tdict)

# Merge gyear & location data
pdf = fastparquet.ParquetFile("RawData/Cleaned/patent_loc_unique_us_0628.parq").to_pandas(["patent", "inv_msa", "gyear"])
cit = cit.merge(pdf.add_prefix("citing_"), left_on = "citing", right_on = "citing_patent").drop("citing_patent",1)
# Cited patent's grant year
cit = cit.merge(pdf[["patent", "gyear"]].add_prefix("cited_"), left_on = "cited", right_on = "cited_patent").drop("cited_patent",1)
# Target patent's location
cit = cit.merge(pdf[["patent", "inv_msa"]].add_prefix("tp_"), left_on = "tp", right_on = "tp_patent").drop("tp_patent",1)
del(pdf)

# Citation occured within 10 years of cited patent's grant date
cit["year_diff"] = cit["citing_gyear"]-cit["cited_gyear"]
print(len(cit))

# Citing inv msa matches tp inv msa
cit["inv_msa_match"] = (cit["tp_inv_msa"] == cit["citing_inv_msa"])

# Is cited patent target
cit["is_tp"] = (cit["tp"] == cit["cited"])

cit.head()

fastparquet.write("DataStore/2018-07-P2/tp_cp_cites_0717.parq", cit, compression="GZIP")

### Summarize info by each target patent

- If tp_cite_match is missing then target patent did not receive any citations; remove from sample

# cit = fastparquet.ParquetFile("DataStore/2018-07-P2/tp_cp_cites_0717.parq").to_pandas()
ts = pd.read_pickle("DataStore/2018-07-P2/targ_samp_0717.pkl")

# Remove non-cited patents from sample: some may have vanished from self-citation cleaning
print(len(ts))

tp_match_5 = cit.loc[(cit["is_tp"] == True) & (cit["year_diff"] <= 5), ["tp", "inv_msa_match"]].groupby(["tp"]).mean()
cp_match_5 = cit.loc[(cit["is_tp"] == False) & (cit["year_diff"] <= 5), ["tp", "inv_msa_match"]].groupby(["tp"]).mean()

tp_match_10 = cit.loc[(cit["is_tp"] == True) & (cit["year_diff"] <= 10), ["tp", "inv_msa_match"]].groupby(["tp"]).mean()
cp_match_10 = cit.loc[(cit["is_tp"] == False) & (cit["year_diff"] <= 10), ["tp", "inv_msa_match"]].groupby(["tp"]).mean()

ts["tp_match_5"] = ts["tp"].map(pd.Series(tp_match_5["inv_msa_match"]))
ts["cp_match_5"] = ts["tp"].map(pd.Series(cp_match_5["inv_msa_match"]))

ts["tp_match_10"] = ts["tp"].map(pd.Series(tp_match_10["inv_msa_match"]))
ts["cp_match_10"] = ts["tp"].map(pd.Series(cp_match_10["inv_msa_match"]))

# Drop all values that have all 4 columns missing; some self-citation removal might have made this empty
ts = ts.dropna(subset=["tp_match_5", "cp_match_5", "tp_match_10", "cp_match_10"], how="all")
print(len(ts))

ts.head()

ts[["tp_gyear", "tp_match_5", "cp_match_5", "tp_match_10", "cp_match_10"]].groupby("tp_gyear").mean()

ts.to_pickle("DataStore/2018-07-P2/targ_samp_0717.pkl")

### Add target's citing patents

- Remove rows where there are no target citations

cit = fastparquet.ParquetFile("DataStore/2018-07-P2/tp_cp_cites_0717.parq").to_pandas()

cit = cit.loc[cit["is_tp"] == True]

for yd in [5,10]:
    c2 = cit.loc[cit["year_diff"] <= yd]

    # In same MSA
    sm = {n: g["citing"].tolist() for n,g in c2.loc[c2["inv_msa_match"] == True, ["tp", "citing"]].groupby(["tp"])}
    dm = {n: g["citing"].tolist() for n,g in c2.loc[c2["inv_msa_match"] == False, ["tp", "citing"]].groupby(["tp"])}

    ts["cite_msa_match_{0}".format(yd)] = ts["tp"].map(sm)
    ts["cite_msa_diff_{0}".format(yd)] = ts["tp"].map(dm)
    
print(len(ts))

# Remove rows where there are no target citations
ts = ts.dropna(subset=['cite_msa_match_5', 'cite_msa_diff_5', 'cite_msa_match_10',
       'cite_msa_diff_10'], how="all")
print(len(ts))

ts.head()

ts.to_pickle("DataStore/2018-07-P2/targ_samp_0717.pkl")
