==> JTHReplication/3b-JTHRegSim-0726.ipynb <==
import scipy as sp
import numpy as np
import pandas as pd
import timeit
import re
import json
import pickle
import fastparquet
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
import datetime
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels.iolib.summary2 as summary2
import logging

logger = logging.getLogger()
logger.setLevel(logging.INFO)
logging.basicConfig(level=logging.INFO)
logger.addHandler(logging.FileHandler('Logs/reg_{0}.log'.format(datetime.datetime.now().\
                                                            strftime("%Y-%m-%d"), 'a')))
print = logging.info
print('good day to you madam fiona')
print('started')
print(datetime.datetime.now())

## Regression: Comparing target's same MSA cites to different MSA cites

### 1. Create new regression sample using only target's similarities

ts = pd.read_pickle("DataStore/2018-07-P2/targ_samp_cite_sim_0717.pkl")

# Get rid of 2010+
ts = ts.loc[ts["tp_gyear"] < 2010]
# Get rid of cp columns
ts = ts[[c for c in ts.columns if "_cp_" not in c]]

# Add year group
def get_year_group(x):
    if x in range(1975,1980):
        yg = "1975-80"
    elif x in range(1980,1985):
        yg = "1980-85"
    elif x in range(1985, 1990):
        yg = "1985-90"
    elif x in range(1990,1995):
        yg = "1990-95"
    elif x in range(1995,2000):
        yg = "1995-00"
    elif x in range(2000,2005):
        yg = "2000-05"
    elif x in range(2005,2010):
        yg = "2005-10"
    elif x in range(2010, 2015):
        yg = "2010-15"
    else:
        yg = np.nan
    return yg

def get_year_group_10(x):
    if x in range(1975,1985):
        yg = "1975-85"
    elif x in range(1985,1995):
        yg = "1985-95"
    elif x in range(1995, 2005):
        yg = "1995-05"
    elif x in range(2005,2015):
        yg = "2005-15"
    else:
        yg = np.nan
    return yg

ts["year_group"] = ts["tp_gyear"].apply(get_year_group_10)

# Scale
# eps = 0.01
# dv_min = 0.7
# def scale_docvecs(x):
#     scaled = ((x+dv_min)/(1+dv_min))*(1-eps)+eps
#     return scaled
# def scale_ldavecs(x):
#     scaled = x*(1-eps)+eps
#     return scaled

sim_cols = [c for c in ts.columns if "sim_" in c]
# for c in sim_cols:
#     if "docvecs" in c:
#         ts[c] = ts[c].apply(scale_docvecs)
#     else:
#         ts[c] = ts[c].apply(scale_ldavecs)

# Normed values
for c in sim_cols:
    ts["norm_{0}".format(c)] = np.nan
    ts.loc[ts[c].notnull(), "norm_{0}".format(c)] = \
    scaler.fit_transform(ts.loc[ts[c].notnull(), c].values.reshape(-1,1))

ts.head()

ts1 = ts[['tp', 'tp_primclass', 'tp_gyear', 'tp_inv_msa','sim_ldavecs_tp_cite_msa_diff_10',
          'sim_docvecs_tp_cite_msa_diff_10', 'norm_sim_ldavecs_tp_cite_msa_diff_10',
          'norm_sim_docvecs_tp_cite_msa_diff_10',
         "year_group", ]]
ts1.columns = ['tp', 'tp_primclass', 'tp_gyear', 'tp_inv_msa', 
       'sim_ldavecs_tp_cite_10','sim_docvecs_tp_cite_10',
               'norm_sim_ldavecs_tp_cite_10','norm_sim_docvecs_tp_cite_10',
               "year_group"]
ts1["inv_msa_match"] = False
ts2 = ts[['tp', 'tp_primclass', 'tp_gyear', 'tp_inv_msa', 'sim_ldavecs_tp_cite_msa_match_10',
        'sim_docvecs_tp_cite_msa_match_10', 'norm_sim_ldavecs_tp_cite_msa_match_10',
        'norm_sim_docvecs_tp_cite_msa_match_10',
        "year_group"]]
ts2.columns = ['tp', 'tp_primclass', 'tp_gyear', 'tp_inv_msa', 
       'sim_ldavecs_tp_cite_10', 'sim_docvecs_tp_cite_10',
               'norm_sim_ldavecs_tp_cite_10','norm_sim_docvecs_tp_cite_10',
              "year_group"]
ts2["inv_msa_match"] = True
ts = ts1.append(ts2, ignore_index = True)
del(ts1, ts2)

pathdir = "DataStore/2018-07-P3/JTHReg0727/"
ts.to_pickle(pathdir+"tp_cite_loc_sim_0727.pkl")

_____
Start here

## Regressions

pathdir = "DataStore/2018-07-P3/JTHReg0727/"
ts = pd.read_pickle(pathdir+"tp_cite_loc_sim_0727.pkl")
ts["tp_primclass"] = ts["tp_primclass"].astype(int)

# # Remove outliers
# print(len(ts))
# ts = ts.loc[ts['norm_sim_docvecs_tp_cite_10'] >= -4]
# print(len(ts))

dm = "norm_sim_docvecs_tp_cite_10"
display(ts.loc[ts["inv_msa_match"] == True, [dm, "year_group"]].groupby("year_group").mean())
display(ts.loc[ts["inv_msa_match"] == False, [dm, "year_group"]].groupby("year_group").mean())

dm = "sim_docvecs_tp_cite_10"
display(ts.loc[ts["inv_msa_match"] == True, [dm, "year_group"]].groupby("year_group").mean())
display(ts.loc[ts["inv_msa_match"] == False, [dm, "year_group"]].groupby("year_group").mean())

r_f = "DataStore/2018-07-P2/Reg0726/reg_model_0726.pkl"
regs = pickle.load(open(r_f, "rb"))

# Similarity regressions
cit_reg = []
for c in ['sim_ldavecs_tp_cite_10', 'sim_docvecs_tp_cite_10']:
    cit_reg.extend([
        "{0} ~ C(inv_msa_match) + C(tp_gyear)".format(c),
        "{0} ~ C(inv_msa_match) + C(tp_primclass)  + C(tp_gyear)".format(c),
        "norm_{0} ~ C(inv_msa_match) + C(tp_gyear)".format(c),
        "norm_{0} ~ C(inv_msa_match) + C(tp_primclass)  + C(tp_gyear)".format(c),
    ])
    regs["JTH Targ"] = pd.Series(cit_reg)
    
    regs["JTH Targ model names"] = pd.Series([
        "Sim LDAVecs - 10 yrs, Year FE",
        "Sim LDAVecs - 10 yrs, PC FE",
        "N Sim LDAVecs - 10 yrs, Year FE",
        "N Sim LDAVecs - 10 yrs, PC FE",
        "Sim DocVecs - 10 yrs, Year FE",
        "Sim DocVecs - 10 yrs, PC FE",
        "N DocVecs - 10 yrs, Year FE",
        "N DocVecs - 10 yrs, PC FE",
    ])
    
    regs["JTH Targ Norm"] = regs["JTH Targ"].loc[[2,3,6,7]]
pickle.dump(regs, open(r_f, "wb"))

info_dict = {'$N$':lambda x: "{0:d}".format(int(x.nobs)),
'Adjusted $R^2$':lambda x: "{:.2f}".format(x.rsquared_adj)}
def get_fit(formula, grouped_data, group_col, cov_type, return_fit = False):
    summ = []
    tables = {}
    
    # Remove missing values used in formula
    col_used = re.findall('\((.*?)\)',formula)
    grouped_data = grouped_data.dropna(how="any", subset=col_used) # Should take care of gyear > 1980 issues
    
    # Group and then get results
    grouped_data = grouped_data.groupby(group_col)
    
    for n,g in grouped_data:
        try:
            fit = smf.ols(formula = formula, data = g, missing="drop").fit(cov_type=cov_type)
            # Get results tables
            tables[n] = fit.summary2().tables
            # Append results
            summ.append(fit)
        except:
            pass
    # Get full results output
    # Dataframe of full results
    res_no_stars = summary2.summary_col(summ, stars = False, \
    model_names = ["{0}".format(n) for n in tables.keys()],\
        info_dict = info_dict).tables[0]
    res_stars = summary2.summary_col(summ, stars = True, \
    model_names = ["{0}".format(n) for n in tables.keys()],\
        info_dict = info_dict).tables[0]
    
    # Get partial results
    # 1. Get relevant variables from index of full results
    regressors = [v for v in res_no_stars.index.unique() if ("sim_" in v) | ("match" in v)]
    # 2. Make sure regressors come last
    regressors = regressors+["Intercept"]
    # 3. Get results with regressors
    part_res_no_stars = summary2.summary_col(summ, stars = False, \
    model_names = ["{0}".format(n) for n in tables.keys()],\
        info_dict = info_dict, regressor_order = regressors).tables[0]
    part_res_stars = summary2.summary_col(summ, stars = True, \
    model_names = ["{0}".format(n) for n in tables.keys()],\
        info_dict = info_dict, regressor_order = regressors).tables[0]
    
    # 4. Get index of where Intercept is and add 2 (to include standard error)
    last_ind = list(part_res_stars.index).index("Intercept")+2
    
    # 5. Get partial results
    part_res_no_stars = pd.concat([part_res_no_stars.iloc[:last_ind], part_res_no_stars.iloc[-2::]])
    part_res_stars = pd.concat([part_res_stars.iloc[:last_ind], part_res_stars.iloc[-2::]])
    
    if return_fit == True:
        return summ, tables, res_no_stars, res_stars, part_res_no_stars, part_res_stars
    else:
        return tables, res_no_stars, res_stars, part_res_no_stars, part_res_stars

print("started")
print(datetime.datetime.now())
pathdir = "DataStore/2018-07-P3/JTHReg0727/"
# 1. Load file
rs = pd.read_pickle(pathdir+"tp_cite_loc_sim_0727.pkl")
print(len(rs))
# 3. Define output
samp_out = {}
formulas = list(regs["JTH Targ"])
formulas_ind = list(regs["JTH Targ"].index)
cov = "HC1"
for i, j in zip(formulas_ind, formulas):
    print(i)
    print(j)
    print(datetime.datetime.now())
    try:
        out = get_fit(j, rs, "year_group", cov, return_fit = False)
        samp_out[i] = {}
        samp_out[i]["model"] = j
        samp_out[i]["tables"] = out[0]
        samp_out[i]["res_no_stars"] = out[1]
        samp_out[i]["res_stars"] = out[2]
        samp_out[i]["part_res_no_stars"] = out[3]
        samp_out[i]["part_res_stars"] = out[4]
    except Exception as e:
        logging.exception("error here")
        pass
    print("finished")
    print(datetime.datetime.now())
pickle.dump(samp_out, open(pathdir+"JTH_tp_cite_loc_res_0726.pkl", "wb"))    

### Check data

pathdir = "DataStore/2018-07-P3/JTHReg0727/"
ts = pd.read_pickle(pathdir+"tp_cite_loc_sim_0727.pkl")
ts["tp_primclass"] = ts["tp_primclass"].astype(int)
ts.loc[(ts["inv_msa_match"] == True), ['sim_ldavecs_tp_cite_10', 'sim_docvecs_tp_cite_10']].describe()

ts.loc[(ts["inv_msa_match"] == False), ['sim_ldavecs_tp_cite_10', 'sim_docvecs_tp_cite_10']].describe()
