==> JTHReplication/3c-JTHReplicationResults-0726.ipynb <==
import scipy as sp
import numpy as np
import pandas as pd
import timeit
import re
import json
import pickle
import fastparquet
import os
os.chdir('/mnt/t48/bighomes-active/sfeng/patentdiffusion/')
seed = 3
import seaborn as sns
import matplotlib
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
import datetime
import statsmodels.api as sm
import statsmodels.formula.api as smf
import statsmodels.iolib.summary2 as summary2
import logging

## 1. Regression Results

pathdir = "DataStore/2018-07-P3/JTHReg0727/"
r_f = "DataStore/2018-07-P2/Reg0726/reg_model_0726.pkl"
regs = pickle.load(open(r_f, "rb"))
res = pickle.load(open(pathdir+"JTH_cite_res_0727.pkl", "rb"))
res_out = pd.DataFrame()

for k in res.keys():
    lks = res[k]["model"].split(" ~ ")[0]

    # Selecting portion of results without intercept
    cdf = res[k]["part_res_stars"].reset_index()
    ic_ind = cdf.loc[cdf["index"] == "Intercept"].index[0]
    # Include N & R^2
    cdf = cdf.iloc[pd.np.r_[0:ic_ind,ic_ind+2:len(cdf)]]
    # Just include R^2
#             cdf = cdf.iloc[pd.np.r_[0:ic_ind,ic_ind+2:len(cdf)-2, len(cdf)-1:len(cdf)]]
    cdf["Model"] = regs["JTH model names"][k]
    cdf["Model Num"] = k
    cdf["LKS"] = lks

    res_out = res_out.append(cdf)
res_out["samp"] = "JTH Rep"
res_out = res_out.replace("C(inv_msa_match)[T.1]", "$I(MSA \, Match)$")
res_out_Rep = res_out

pathdir = "DataStore/2018-07-P3/JTHReg0727/"
r_f = "DataStore/2018-07-P2/Reg0726/reg_model_0726.pkl"
regs = pickle.load(open(r_f, "rb"))
res = pickle.load(open(pathdir+"JTH_tp_cite_loc_res_0726.pkl", "rb"))
res_out = pd.DataFrame()

for k in res.keys():
    lks = res[k]["model"].split(" ~ ")[0]

    # Selecting portion of results without intercept
    cdf = res[k]["part_res_stars"].reset_index()
    ic_ind = cdf.loc[cdf["index"] == "Intercept"].index[0]
    # Include N & R^2
    cdf = cdf.iloc[pd.np.r_[0:ic_ind,ic_ind+2:len(cdf)]]
    # Just include R^2
#             cdf = cdf.iloc[pd.np.r_[0:ic_ind,ic_ind+2:len(cdf)-2, len(cdf)-1:len(cdf)]]
    cdf["Model"] = regs["JTH Targ model names"][k]
    cdf["Model Num"] = k
    cdf["LKS"] = lks

    res_out = res_out.append(cdf)
res_out = res_out.reset_index(drop=True)
res_out = res_out.replace("C(inv_msa_match)[T.True]", "$I(MSA \, Match)$")
res_out["samp"] = "JTH Targ Sim"
res_out_cs = res_out

r = pd.concat([res_out_Rep, res_out_cs],axis=0)

r.to_csv(pathdir+"JTH_res_out_0726.csv")

r

______
## 2. Replication Tables

### 2.1 JTH Citation Extension

ts = pd.read_pickle("DataStore/2018-07-P2/targ_samp_cite_sim_0717.pkl")
# Get rid of 2010+
ts = ts.loc[ts["tp_gyear"] < 2010]

# Add year group
def get_year_group(x):
    if x in range(1975,1980):
        yg = "1975-80"
    elif x in range(1980,1985):
        yg = "1980-85"
    elif x in range(1985, 1990):
        yg = "1985-90"
    elif x in range(1990,1995):
        yg = "1990-95"
    elif x in range(1995,2000):
        yg = "1995-00"
    elif x in range(2000,2005):
        yg = "2000-05"
    elif x in range(2005,2010):
        yg = "2005-10"
    elif x in range(2010, 2015):
        yg = "2010-15"
    else:
        yg = np.nan
    return yg

def get_year_group_10(x):
    if x in range(1975,1985):
        yg = "1975-85"
    elif x in range(1985,1995):
        yg = "1985-95"
    elif x in range(1995, 2005):
        yg = "1995-05"
    elif x in range(2005,2015):
        yg = "2005-15"
    else:
        yg = np.nan
    return yg

ts["year_group"] = ts["tp_gyear"].apply(get_year_group_10)

sim_cols = [c for c in ts.columns if "sim_" in c]

# Normed values
for c in sim_cols:
    ts["norm_{0}".format(c)] = np.nan
    ts.loc[ts[c].notnull(), "norm_{0}".format(c)] = \
    scaler.fit_transform(ts.loc[ts[c].notnull(), c].values.reshape(-1,1))
    
ts2 = ts.copy()

ts = ts2.copy()
# Remove outliers
pg = ts[["tp_match_10", "cp_match_10", "year_group"]].groupby("year_group")
summ = pg.mean()*100
for yd in [10]:
    summ["$t$-stat".format(yd)] = [sp.stats.ttest_ind(g["tp_match_{0}".format(yd)],\
                                                            g["cp_match_{0}".format(yd)], 
                                                            equal_var = False, nan_policy="omit")[0]
                                        for n,g in pg] 
    summ["$p$-value".format(yd)] = [sp.stats.ttest_ind(g["tp_match_{0}".format(yd)],\
                                                            g["cp_match_{0}".format(yd)], 
                                                            equal_var = False, nan_policy="omit")[1]
                                        for n,g in pg]
    summ["$N$ obs"] = [len(g) for n,g in pg]
    summ["Difference".format(yd)] = summ["tp_match_{0}".format(yd)]-summ["cp_match_{0}".format(yd)]

summ = np.round(summ.reset_index(),2)

summ = summ[['year_group', 'tp_match_10', 'cp_match_10','Difference', '$t$-stat', '$p$-value', '$N$ obs',]]
summ.columns = ['Grant Year', 'Pct Targ Cites in $MSA_T$', 'Pct Control Cites in $MSA_T$','Diff', '$t$-stat', '$p$-value', '$N$ obs',]
# Add proportional difference
# summ["Prop Diff"] = np.round(summ["Difference"]/summ['cp_match_10'],2)
# summ = summ[['year_group', 'tp_match_10', 'cp_match_10','Difference', "Prop Diff", '$t$-stat', '$p$-value', '$N$ obs',]]
# summ.columns = ['Grant Year', 'Pct Targ Cites in $MSA_T$', 'Pct Control Cites in $MSA_T$','Abs Diff',"Prop Diff", '$t$-stat', '$p$-value', '$N$ obs',]
# Add proportional difference

summ1 = summ
# summ1 = summ.reset_index()

# Erase last row!
display(summ)
print(summ.to_latex(index=False,escape=False,column_format="lccccccc"))

### 2.2 JTH Citation Similarity

pg = ts[['sim_ldavecs_tp_cite_msa_match_10', 'sim_ldavecs_tp_cite_msa_diff_10',
         'sim_docvecs_tp_cite_msa_match_10', 'sim_docvecs_tp_cite_msa_diff_10',
         "year_group"]].groupby("year_group")

summ = pg.mean()
for dm in ["ldavecs", "docvecs"]:
    x1 = 'sim_{0}_tp_cite_msa_match_10'.format(dm)
    x2 = 'sim_{0}_tp_cite_msa_diff_10'.format(dm)
    summ["$t$-stat ({0})".format(dm)] = [sp.stats.ttest_ind(g[x1],\
                                                            g[x2], 
                                                            equal_var = False, nan_policy="omit")[0]
                                        for n,g in pg] 
    summ["$p$-value ({0})".format(dm)] = [sp.stats.ttest_ind(g[x1],\
                                                            g[x2], 
                                                            equal_var = False, nan_policy="omit")[1]
                                        for n,g in pg]
#     summ["$N$ obs"] = [len(g) for n,g in pg]
    summ["Diff {0}".format(dm)] = summ[x1]-summ[x2]

summ = np.round(summ,2)
summ4 = summ.reset_index()

v_dict = dict(zip(['year_group', 'sim_ldavecs_tp_cite_msa_match_10',
       'sim_ldavecs_tp_cite_msa_diff_10','Diff ldavecs','$t$-stat (ldavecs)',
       '$p$-value (ldavecs)', 'sim_docvecs_tp_cite_msa_match_10',
       'sim_docvecs_tp_cite_msa_diff_10', 'Diff docvecs','$t$-stat (docvecs)',
       '$p$-value (docvecs)',],
     ["Grant Year", r"$\overline{sim_{LDA}(T,j|j\, in \,MSA_T)}$", 
     "$\\overline{sim_{LDA}(T,j|j \, not \, in \,MSA_T)}$", "Diff", "$t$-stat", "$p$-value",
     "$\\overline{sim_{DV}(T,j|j\, in \,MSA_T)}$",
    "$\\overline{sim_{DV}(T,j|j \, not \, in \,MSA_T)}$", "Diff","$t$-stat", "$p$-value", ]))

summ5 = summ4[['year_group', 'sim_ldavecs_tp_cite_msa_match_10',
       'sim_ldavecs_tp_cite_msa_diff_10','Diff ldavecs','$t$-stat (ldavecs)',
       '$p$-value (ldavecs)',]]
summ5 = summ5.rename(columns=v_dict)
summ6 = summ4[['year_group', 'sim_docvecs_tp_cite_msa_match_10',
       'sim_docvecs_tp_cite_msa_diff_10', 'Diff docvecs','$t$-stat (docvecs)',
       '$p$-value (docvecs)',]]
summ6 = summ6.rename(columns=v_dict)

print("JTHOwnCiteSim")
display(summ5)
print(summ5.to_latex(index=False,escape=False, column_format="lccccc"))
display(summ6)
print(summ6.to_latex(index=False,escape=False, column_format="lccccc"))

ts['sim_docvecs_tp_cite_msa_match_10'].describe()
